requirements.txt
bandit==1.7.5
coverage==7.3.2
flake8==6.1.0
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
ruamel.yaml==0.18.5
stringcase==1.2.0
yahoo-oauth==2.0
yfpy
PyYAML
pandas
streamlit